Long & instance 1000

+++ Instance.h
@@ -7,9 +7,9 @@
 
 class Instance {
  private:
-  std::vector<int> durations;
-  std::vector<int> weights;
-  std::vector<int> deadlines; 
+  std::vector<long> durations;
+  std::vector<long> weights;
+  std::vector<long> deadlines; 
   
   int id;
   int instance_size;
@@ -19,9 +19,9 @@ class Instance {
   (
    int id,
    int instance_size,
-   std::vector<int> durations,
-   std::vector<int> weights,
-   std::vector<int> deadlines
+   std::vector<long> durations,
+   std::vector<long> weights,
+   std::vector<long> deadlines
   ):
    id(id),
    instance_size(instance_size),
@@ -32,13 +32,13 @@ class Instance {
 
   int get_id() { return id; }
   int get_instance_size() { return instance_size; }
-  int get_duration(int i) { return durations[i]; }
-  int get_weight(int i) { return weights[i]; }
-  int get_deadline(int i) { return deadlines[i]; }
+  long get_duration(int i) { return durations[i]; }
+  long get_weight(int i) { return weights[i]; }
+  long get_deadline(int i) { return deadlines[i]; }
   
-  std::vector<int> get_durations() { return durations; }
-  std::vector<int> get_weights() { return weights; }
-  std::vector<int> get_deadlines() { return deadlines; }
+  std::vector<long> get_durations() { return durations; }
+  std::vector<long> get_weights() { return weights; }
+  std::vector<long> get_deadlines() { return deadlines; }
 
   void display();
 };
